requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
httpx>=0.25.0
pydantic>=2.8.2
camel-ai>=0.2.0
python-dotenv>=1.0.0